requirements.txt
# Langchain
langchain
langchain-core
langchain-community
langchain-openai
langchain_postgres
langsmith
langgraph
duckduckgo-search
bs4
# # Lamma Cloud
# llama-index-core
# llama-parse
# llama-index-readers-file
# llama-index-embeddings-openai

# # huggingface
# transformers
# torch
# datasets

# Others
python-dotenv
streamlit
python-docx